@@ -17,7 +17,6 @@ if ( ! defined( 'WPINC' ) ) {
 
 
 class GravityView_Edit_Entry {
-
     /**
      * @var string
      */
@@ -95,6 +94,11 @@ class GravityView_Edit_Entry {
 		add_filter( 'gravityview/field/is_visible', array( $this, 'maybe_not_visible' ), 10, 3 );
 
 		add_filter( 'gravityview/api/reserved_query_args', array( $this, 'add_reserved_arg' ) );
+
+		add_filter( 'gform_notification_events', array( $this, 'add_edit_notification_events' ), 10, 2 );
+
+		add_action( 'gravityview/edit_entry/after_update', array( $this, 'trigger_notifications' ), 10, 3 );
+
     }
 
 	/**
@@ -348,7 +352,7 @@ class GravityView_Edit_Entry {
     /**
      * checks if user has permissions to edit a specific entry
      *
-     * Needs to be used combined with GravityView_Edit_Entry::user_can_edit_entry for maximum security!!
+     * Needs to be used combined with GravityView_Edit_Entry_Render::user_can_edit_entry for maximum security!!
      *
      * @param  array|\WP_Error $entry Gravity Forms entry array or WP_Error if the entry wasn't found.
      * @param \GV\View|int    $view ID of the view you want to check visibility against {@since 1.9.2}. Required since 2.0.
@@ -399,7 +403,7 @@ class GravityView_Edit_Entry {
             $current_user = wp_get_current_user();
 
             // User edit is disabled
-            if ( empty( $user_edit ) ) {
+            if ( $view_id && empty( $user_edit ) ) {
 
                 gravityview()->log->debug( 'User Edit is disabled. Returning false.' );
 
@@ -455,7 +459,35 @@ class GravityView_Edit_Entry {
 
 		GFForms::delete_file();
 	}
-} // end class
 
-// add_action( 'plugins_loaded', array('GravityView_Edit_Entry', 'getInstance'), 6 );
+	/**
+	 * Triggers notifications.
+	 *
+	 * @since 2.32.0
+	 *
+	 * @param array                         $form              The form object.
+	 * @param int                           $entry_id          The entry ID.
+	 * @param GravityView_Edit_Entry_Render $edit_entry_render The edit entry render class instance.
+	 */
+	public function trigger_notifications( $form, $entry_id, $edit_entry_render ) {
+		GravityView_Notifications::send_notifications( (int) $entry_id, 'gravityview/edit_entry/after_update', $edit_entry_render->entry );
+	}
+
+	/**
+	 * Adds the notification event.
+	 *
+	 * @since 2.32.0
+	 *
+	 * @param array $notification_events Existing notification events.
+	 * @param array $form                The form object.
+	 *
+	 * @return array
+	 */
+	public function add_edit_notification_events( $notification_events = [], $form = [] ) {
+		$notification_events['gravityview/edit_entry/after_update'] = 'GravityView - ' . esc_html_x( 'Entry is updated', 'The title for an event in a notifications drop down list.', 'gk-gravityview' );
+
+		return $notification_events;
+	}
+}
+
 GravityView_Edit_Entry::getInstance();
