@@ -132,10 +132,21 @@ abstract class GravityView_Plugin_and_Theme_Hooks {
 	private function maybe_add_hooks() {
 
 		if ( ! $this->is_active() ) {
+
+			$this->add_inactive_hooks();
+
 			return;
 		}
 
 		$this->add_hooks();
+	}
+
+	/**
+	 * Add hooks when the plugin / theme is active.
+	 *
+	 * @since 2.26
+	 */
+	protected function add_inactive_hooks(): void {
 	}
 
 	/**
@@ -178,8 +189,7 @@ abstract class GravityView_Plugin_and_Theme_Hooks {
 	 * @return array Array of features associated with a functional area of the edit screen, merged with existing values
 	 */
 	public function merge_post_type_support( $supports = array(), $is_hierarchical = false ) {
-		$supports = array_merge( $this->post_type_support, $supports );
-		return $supports;
+		return array_merge( $this->post_type_support, $supports );
 	}
 
 	/**
