@@ -27,7 +27,8 @@ class GravityView_Theme_Hooks_Elegant_Themes extends GravityView_Plugin_and_Them
 	function add_hooks() {
 		parent::add_hooks();
 
-		add_action( 'admin_init', array( $this, 'add_hooks_admin_init' ), 1 );
+		add_action( 'admin_init', [ $this, 'add_hooks_admin_init' ], 1 );
+		add_filter( 'et_builder_enable_jquery_body', [ $this, 'edit_entry_jquery_compat_fix' ] );
 	}
 
 	/**
@@ -98,6 +99,22 @@ class GravityView_Theme_Hooks_Elegant_Themes extends GravityView_Plugin_and_Them
 		}
 
 		return $shortcodes;
+	}
+
+	/**
+	 * Prevents Divi from "faking" jQuery on the Edit Entry screen.
+	 * This breaks some functionality, such as the Gravity Forms Signature Add-On.
+	 *
+	 * @used-by `et_builder_enable_jquery_body` filter
+	 *
+	 * @since 2.31.0
+	 *
+	 * @param bool $enable_jquery_compat
+	 *
+	 * @return bool
+	 */
+	public function edit_entry_jquery_compat_fix( $enable_jquery_compat ) {
+		return gravityview()->request->is_edit_entry() ? false : $enable_jquery_compat;
 	}
 }
 
