@@ -37,7 +37,13 @@ class GravityView_Plugin_Hooks_Gravity_Flow extends GravityView_Plugin_and_Theme
 
 		add_filter( 'gravityview/admin/available_fields', array( $this, 'maybe_add_non_default_fields' ), 10, 3 );
 
-		add_filter( 'gravityview/adv_filter/field_filters', array( $this, 'maybe_add_non_default_filter_fields' ), 10, 2 );
+		if ( defined( 'GRAVITYKIT_ADVANCED_FILTERING_VERSION' ) ) {
+			add_filter(
+				version_compare( GRAVITYKIT_ADVANCED_FILTERING_VERSION, '3.0.0', '<' )
+					? 'gravityview/adv_filter/field_filters'
+					: 'gk/query-filters/field-filters',
+				[ $this, 'maybe_add_non_default_filter_fields' ], 10, 2 );
+		}
 
 		add_action( 'gravityflow_post_process_workflow', array( $this, 'clear_cache_after_workflow' ), 10, 4 );
 
