@@ -46,4 +46,14 @@ $output = apply_filters( 'gravityview_entry_link', GravityView_API::replace_vari
 
 $href = GravityView_Edit_Entry::get_edit_link( $entry, $gravityview->view->ID, $post ? $post->ID : null );
 
-echo gravityview_get_link( $href, $output, $link_atts );
+$output = gravityview_get_link( $href, $output, $link_atts );
+
+/**
+ * Modify the link HTML (here for backward compatibility).
+ *
+ * @param string $output HTML output of the link
+ * @param string $href URL of the link
+ * @param array  $entry The GF entry array
+ * @param  array $field_settings Settings for the particular GV field
+ */
+echo apply_filters( 'gravityview_field_entry_link', $output, $href, $entry, $field_settings );
