@@ -15,6 +15,7 @@ if ( ! isset( $gravityview ) || empty( $gravityview->template ) ) {
 $field         = $gravityview->field;
 $display_value = $gravityview->display_value;
 $input_id      = gravityview_get_input_id_from_id( $field->ID );
+$form_id 	   = $gravityview->view->form->ID;
 
 // Used in filters below.
 $return_true = function () {
@@ -142,7 +143,8 @@ switch ( $gravityview->field->field->inputType ) {
 			return;
 		}
 
-		$choices = $field->field->choices;
+		$choices = GravityView_Field_Survey::get_field_choices( $form_id, $field->ID );
+
 		$choice_values = wp_list_pluck( $choices, 'value', $gravityview->value );
 		$starred_index = array_search( $gravityview->value, $choice_values );
 		$star_a11y_label = $starred_index !== false
